+++ src/init-action.ts
@@ -13,17 +13,12 @@ import {
   StatusReportBase,
 } from "./actions-util";
 import { getGitHubVersion } from "./api-client";
-import {
-  CodeQL,
-  CODEQL_VERSION_NEW_TRACING,
-  enrichEnvironment,
-} from "./codeql";
+import { CodeQL } from "./codeql";
 import * as configUtils from "./config-utils";
 import { Feature, Features } from "./feature-flags";
 import {
   initCodeQL,
   initConfig,
-  injectWindowsTracer,
   installPythonDeps,
   runInit,
   ToolsSource,
@@ -35,7 +30,6 @@ import { getTotalCacheSize } from "./trap-caching";
 import {
   checkForTimeout,
   checkGitHubVersionInRange,
-  codeQlVersionAbove,
   DEFAULT_DEBUG_ARTIFACT_NAME,
   DEFAULT_DEBUG_DATABASE_NAME,
   getMemoryFlagValue,
@@ -219,7 +213,7 @@ async function run() {
   );
 
   try {
-    const workflowErrors = await validateWorkflow();
+    const workflowErrors = await validateWorkflow(logger);
 
     if (
       !(await sendStatusReport(
@@ -252,7 +246,6 @@ async function run() {
     toolsDownloadDurationMs = initCodeQLResult.toolsDownloadDurationMs;
     toolsVersion = initCodeQLResult.toolsVersion;
     toolsSource = initCodeQLResult.toolsSource;
-    await enrichEnvironment(codeql);
 
     config = await initConfig(
       getOptionalInput("languages"),
@@ -356,19 +349,6 @@ async function run() {
     if (tracerConfig !== undefined) {
       for (const [key, value] of Object.entries(tracerConfig.env)) {
         core.exportVariable(key, value);
-      }
-
-      if (
-        process.platform === "win32" &&
-        !(await codeQlVersionAbove(codeql, CODEQL_VERSION_NEW_TRACING))
-      ) {
-        await injectWindowsTracer(
-          "Runner.Worker.exe",
-          undefined,
-          config,
-          codeql,
-          tracerConfig
-        );
       }
     }